Atualização com solução e simulação

+++ Trabalho_PPOO/Ingresso.java
@@ -1,36 +1,56 @@
 /**
  * Classe que representa um ingresso de determinada sessao que pertence ao cinema.
- * COLOCAR O QUE FAZ (ex: permite realizar a compra do ingresso)
  */
 
 public class Ingresso {
-    //Atributos
     private int numero;
-    private int valor;
+    private double valor;
     private String tipoIngresso;
 
-    //Construtor
-    public Ingresso(String filme, int numero, int valor, String tipoIngresso){
+    /**
+     * Construtor da classe Ingresso.
+     * @param filme - string que armazena o nome do filme
+     * @param numero - inteiro que armazena o número do ingresso
+     * @param valor - double que armazena o valor do ingresso
+     * @param tipoIngresso - string que armazena o tipo de ingresso
+     */
+    public Ingresso(String filme, int numero, double valor, String tipoIngresso){
         this.numero = numero;
         this.valor = valor;
         this.tipoIngresso = tipoIngresso;
     }
-
-    //Getters
+    /**
+     * Getter do número do ingresso.
+     * @return int - codigo do ingresso.
+     */
     public int getNumero(){
         return numero;
     }
-    public int getValor(){
+    /**
+     * Getter do valor do ingresso.
+     * @return double - valor do ingresso
+     */
+    public double getValor(){
         return valor;
     }
+    /**
+     * Getter do tipo de ingresso.
+     * @return String - tipo de ingresso.
+     */
     public String getTipoIngresso(){
         return tipoIngresso;
     }
-    
-    //Setters
+    /**
+     * Setter do numero do ingresso.
+     * @param novoNumero - inteiro que armazena o código do ingresso.
+     */
     public void setNumero(int novoNumero){
         numero=novoNumero;
     }
+    /**
+     * Setter do valor do ingresso.
+     * @param novoValor - inteiro que armazena o valor do ingresso.
+     */
     public void setValor(int novoValor){
         valor=novoValor;
     }
